Move announcement.go to handlers dir
